// src/api/generated/models/PassStatusDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PassStatusDto = {
    /**
     * Whether the token ID is valid
     */
    valid: boolean;
    /**
     * The token ID that was validated
     */
    tokenId: string;
};